+++ source/DruidXBasicAuthExtension/src/main/java/com/amazon/solutions/druid/xbasicauth/XBasicHTTPFilter.java
@@ -26,15 +26,6 @@ import java.util.List;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequestWrapper;
 import org.apache.druid.java.util.common.logger.Logger;
-import org.pac4j.core.config.Config;
-import org.pac4j.core.context.J2EContext;
-import org.pac4j.core.context.session.SessionStore;
-import org.pac4j.core.engine.CallbackLogic;
-import org.pac4j.core.engine.DefaultCallbackLogic;
-import org.pac4j.core.engine.DefaultSecurityLogic;
-import org.pac4j.core.engine.SecurityLogic;
-import org.pac4j.core.http.adapter.HttpActionAdapter;
-import org.pac4j.core.profile.CommonProfile;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
@@ -46,39 +37,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class BasicHTTPFilter implements Filter {
-    private static final Logger logger = new Logger(BasicHTTPFilter.class);
+public class XBasicHTTPFilter implements Filter {
+    private static final Logger logger = new Logger(XBasicHTTPFilter.class);
 
-    private final Config pac4jConfig;
-    private final OidcConfig oidcConfig;
-    private final SecurityLogic<CommonProfile, J2EContext> securityLogic;
-    private final CallbackLogic<CommonProfile, J2EContext> callbackLogic;
-    private final SessionStore<J2EContext> sessionStore;
-    private static final HttpActionAdapter<CommonProfile, J2EContext> NOOP_HTTP_ACTION_ADAPTER = (int code,
-            J2EContext ctx) -> null;
-
-    private final String name;
-    private final String authorizerName;
-
-    public BasicHTTPFilter(String name, String authorizerName, Config pac4jConfig, OidcConfig oidcConfig,
-                           String cookiePassphrase) {
-        this(name, authorizerName, pac4jConfig, oidcConfig, cookiePassphrase, new DefaultSecurityLogic<>(),
-                new DefaultCallbackLogic<>());
-    }
-
-    public BasicHTTPFilter(String name, String authorizerName, Config pac4jConfig, OidcConfig oidcConfig,
-                           String cookiePassphrase, SecurityLogic<CommonProfile, J2EContext> securityLogic,
-                           CallbackLogic<CommonProfile, J2EContext> callbackLogic) {
-        this.pac4jConfig = pac4jConfig;
-        this.oidcConfig = oidcConfig;
-        this.securityLogic = securityLogic;
-        this.callbackLogic = callbackLogic;
-
-        this.name = name;
-        this.authorizerName = authorizerName;
-
-        this.sessionStore = new OidcSessionStore<>(cookiePassphrase);
-    }
+    public XBasicHTTPFilter() {}
 
     @Override
     public void init(FilterConfig filterConfig) {
@@ -90,7 +52,7 @@ public class BasicHTTPFilter implements Filter {
             throws IOException, ServletException {
         HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
         HttpServletResponse httpServletResponse = (HttpServletResponse) servletResponse;
-        String xBasicHeader = httpServletRequest.getHeader("x-authorization");
+        String xBasicHeader = httpServletRequest.getHeader("x-auth");
 
         HeaderMapRequestWrapper requestWrapper = new HeaderMapRequestWrapper(httpServletRequest);
