+++ source/DruidXBasicAuthExtension/src/test/java/com/amazon/solutions/druid/xbasicauth/XBasicHTTPFilterTest.java
@@ -1,0 +1,61 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.amazon.solutions.druid.xbasicauth;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import java.io.IOException;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.junit.Before;
+import org.junit.Test;
+
+public class XBasicHTTPFilterTest {
+    private XBasicHTTPFilter filter;
+    private HttpServletRequest request;
+    private HttpServletResponse response;
+
+    @Before
+    public void setup() {
+        request = mock(HttpServletRequest.class);
+        response = mock(HttpServletResponse.class);
+
+        filter = new XBasicHTTPFilter();
+    }
+
+    @Test
+    public void canProcessAuthenticationRequest() throws IOException, ServletException {
+        // arrange
+        when(request.getHeader("x-auth")).thenReturn("Basic test");
+        FilterChain filterChain = mock(FilterChain.class);
+
+        // act
+        filter.doFilter(request, response, filterChain);
+
+        // assert
+        verify(filterChain).doFilter(request, response);
+    }
+}
